backend/requirements.txt
uvicorn[standard]==0.27.1
fastapi[standard]==0.110.0
sqlalchemy[asyncio]==2.0.28
asyncpg==0.29.0
pydantic-settings==2.2.1
pydantic==2.6.3
alembic==1.13.1
arq==0.25.0
redis==5.0.3
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
python-multipart==0.0.9
